Added lambda execution to interpreter. When inside a function, replace type annotations on lambdas and embeds with bound type parameters. I.e.:   ```   fn make_lambda<T>(list: List<T>) -> T -> List<T>     # This lambda is parsed as type T -> List<T>, but it really should     # be @T -> List<@T> (where @ means that it is a parameter type).     # All embeds and lambda literals with any types with the same name     # as a type parameter in the surrounding fn have those types     # replaced with the actual type parameters.     { (x: T) -> x::list }   end   ```
Refactored function resolution so that the code could be shared between
the semantic analyzer and interpreter.

All values passed to an embed call are wrapped in the `Value` class
instead of passed as raw Ruby values. This ensures that the value's type
information is retained when added to a List. For backwards
compatibility, `method_missing?` is used to delegate methods to the raw
value underneath.

Allow lambdas as terminal arguments to a fn to be placed outside the
parens, i.e. `foo(a, b) { ... }`

Allow AST terms to be actual syntax trees themselves instead of just
symbols or literals. This change was necessary to allow lambdas created
from means other than a literal (like as the return value of a fn) to be
called immediately instead of saved in a variable. I.e.:
  ```
  fn foo() -> (Int, Int) -> Int # foo return a lambda
  foo()(1) # Calls foo then calls the lambda returned from foo

  # AST looks like:
  (
    (:foo, []) # This is the term, an AST itself
    (1 : Int) # These are the children
    (2 : Int)
  )
  ```

Parse out comments starting with '#'

Added a native pipe operator that is rewritten by semantic. It passes a
value to the following function as the first argument, like Elixir:
  ```
  [1, 2, 3] |> map { ... } |> reverse

  # is the same as

  reverse(map([1, 2, 3], { ... }))
  ```

Added some basic list function to the kernel.