Replace type bindings in inferred untyped lambda.
If a candidate is found for an untyped lambda that has a type parameter
as its return value, the arg types for the candidate are used when
visiting lambda to determine an inferred return type. That return type
is then used to bind to all instances of the type parameter in the
original candidate type.

This works only for simple parameter types in lambda's return type, so
it won't work for complex parameter types like (Int | @T). This should
be changed in the future. Actually, the entire type inference/type
binding/type checking stuff should be rewritten using a real algorithm.