Overhauled compiler errors and error message.
Now the Lexer, Parser, Semantic, and Scope classes each have their own
error classes. The location of an error in source code is tracked with
Location instances. Each token has a location, and each AST has a token.
So any error during compilation can be tracked to a specific location in
the source. This is used to print a prettier error message in the REPL
included the source code with the location highlighted. This should be
extracted to a general error message class that can be used during
regular compilation as well, not just in the REPL.

Started on a more description spec.
Started on a struct type.
In list module, change reduce to take a `(acc, x) -> ?` lambda instead
of `(x, acc) -> ?`. This follows standard practice in other functional
languages.
Added `reduce_right` and list concatentation to list module.