Added record parsing, analyzing, and interpreting.
Went with the term "record" instead of "struct", since this matches the
term used by other functional languages (Haskell, Elm, OCaml, F#, etc.).
The term "struct" was originally borrowed from Elixir. However, Elixir's
structs are a little different than the record types implemented in this
language. Specifically, Elixir's structs are nominal types; they can
only match other structs with the same name, even if both structs have
the same fields, i.e. `%Foo{}` only matches another `%Foo{}`. This is in
contrast to the planned uses for records in this language; records can
match with -- and be subtypes or supertypes of -- other records
structurally. Vanilla records have no names, only fields. Named types
will be implemented in the future to allow for nominal typing of
records, so the goal is to support both nominal and structural typing
for records. Hence, the move away from the term "struct".