Fixed module and function resolution bug.
This bug happens when `Foo.Bar.baz()` could find and execute
`Foo.baz()` because module and fn resolution is allowed to check the
outer scope for term. Now, the outer scope is only checked for the first
module in a path (`Foo` in `Foo.Bar`) or for a bare fn call (just
`foo`).

Added notes for a newtype feature to be implemented.