Implemented case in the interpreter. Fixed `List`. The compiler and REPL should now be back to stable. Case expressions are implemented in the compiler and interpreter.
Reworked how frames are implemented in the interpreter. Now a frame can
have a parent frame, which expands into frame trees. This was necessary
to handle the complex scoping issues between variables in case pattern
heads and variables in case branches.

Don't throw a compilation error if a function body returns a subtype of
the annotated return type (instead of the exact same type).

Properly handle parameter types in case patterns.

Improved how untyped lambdas have their types inferred. Now, the
lambda's body is repeatedly visited by the type checked to infer the
return type. Any type parameters in this return type are then used to
rebind type parameters in the original lambda type with more general arg
types. These arg types are then used again when visiting the lambda to
infer a new return type, and so on until no more general types are found
or the lambda fails to type check.

Rewrote much of the List module to take advantage of the new case
expressions. Completely removed the concept of an `Empty` type.

Various bugs and tweaks.